requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
msal==1.24.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
numpy==1.24.3
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gunicorn==21.2.0
google-cloud-firestore==2.14.0
firebase-admin==6.4.0